admin-backend/pyproject.toml
[tool.poetry]
name = "admin-backend"
version = "0.1.0"
description = "Smart TG admin backend service"
authors = ["Smart TG DevTeam"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
pydantic = "^2.9.0"
pydantic-settings = "^2.4.1"
sqlalchemy = "^2.0.36"
alembic = "^1.13.3"
passlib = "^1.7.4"
bcrypt = "^4.2.0"
python-jose = "^3.3.0"
redis = "^5.2.0"
python-multipart = "^0.0.7"
httpx = "^0.27.2"
email-validator = "^2.2.0"
slowapi = "^0.1.9"
pyrogram = "^2.0.106"
pyyaml = "^6.0.3"
openai = "^2.8.0"
cryptography = "^41.0.0"
paramiko = "^3.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.2"
black = "^24.10.0"
ruff = "^0.6.9"
pytest-cov = "^7.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

